Move middleware and helpers to separate modules
